ai-employee-system/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
anthropic
win10toast
pyyaml